Turn Controller and Enemy Move Skeletons done
TurnOrderController - Got some really basic turn switching logic in, but that's really all it needs. If we want to add animations of something like a turn wheel, this is probably where I'd do that.

EnemyMove - I did everything I could in this script except for the actual turn, which mostly included declaring variables and the receiving end of the TOC. Also didn't do the animations & stuff like that, but I also have no sprites for that yet, so it'll have to wait.