Work dump
Been working on a lot of scripts. Currently working on getting the EnemyMove to do what I want it to, specifically the chasing part. I have two options:
1. I could make the enemies chase the player to the ends of the earth. This would lead to a higher level of difficulty, as the player would be harder pressed to find a safe spot. However, it's a lot of work: so much, in fact, that it was basically my whole final project last year. And yes, I could copy and paste it, but it would take a while to configure.
2. I could make the enemies give up the chase after they lose sight of their home location, if they have one to return to. Easier solution, and a lot less work.
Currently, I'm going to go with option 2, and then if I have time at the end, I can update it. Frankly, I'm not sure how much of my ideas I can finish, especially with IT being the way they are, so I'm trying to just get some mechanics working so I can touch on everything.