Light controls should be functional
FOR USE: Make sure that the Lights Controller prefab is in the scene, and group all of the Light GameObjects as children under a GameObject called "LightsGameObjects". Also, make sure that the GlobalLight is the first (favorite) child.

The FlickeringLights script was also added. FlickerIntensityRange changes how much the light changes every time it flickers (Default: 0.15). FlickerTimerMax changes how often it flickers (Default 0.1)