Save Push
Been working hard, the animators for the Gem Pickup Canvas are being stubborn, tho. The fragment image is showing up too early, but everything else is (to my knowledge) functional.

Also set up some frameworks for changing the scene when the gem is picked up, because that might be fun to do if I continue the project